Proyecto3_Actualizo, nombres en torneo se repiten

+++ Proyectos/proyecto3/GUI/InterfazPrueba/src/sistema/Peleador.java
@@ -20,7 +20,7 @@ public class Peleador {
     }
 
     public double getProbabilidad(Peleador oponente) {
-        return habilidad / (habilidad + oponente.getHabilidad());
+        return habilidad / (double)(habilidad + oponente.getHabilidad());
     }
 
 }
